chore: Avoid warnings and errors when using different gRPC versions
